+++ internal/handler/proxy/middleware/errorhandler/mocks/error_handler.go
