+++ internal/handler/proxy/middleware/errorhandler/options.go
