+++ internal/handler/proxy/middleware/prometheus/defaults.go
