+++ internal/handler/proxy/middleware/prometheus/handler_test.go
