+++ internal/handler/proxy/middleware/recovery/handler.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/rs/zerolog"
 
-	"github.com/dadrus/heimdall/internal/handler/proxy/middlewares/errorhandler"
+	"github.com/dadrus/heimdall/internal/handler/proxy/middleware/errorhandler"
 	"github.com/dadrus/heimdall/internal/heimdall"
 	"github.com/dadrus/heimdall/internal/x/errorchain"
 	"github.com/dadrus/heimdall/internal/x/stringx"
