+++ internal/handler/proxy/service.go
@@ -34,14 +34,14 @@ import (
 
 	"github.com/dadrus/heimdall/internal/cache"
 	"github.com/dadrus/heimdall/internal/config"
-	"github.com/dadrus/heimdall/internal/handler/proxy/middlewares/accesslog"
-	cachemiddleware "github.com/dadrus/heimdall/internal/handler/proxy/middlewares/cache"
-	"github.com/dadrus/heimdall/internal/handler/proxy/middlewares/dump"
-	"github.com/dadrus/heimdall/internal/handler/proxy/middlewares/errorhandler"
-	"github.com/dadrus/heimdall/internal/handler/proxy/middlewares/logger"
-	prometheusmiddleware "github.com/dadrus/heimdall/internal/handler/proxy/middlewares/prometheus"
-	"github.com/dadrus/heimdall/internal/handler/proxy/middlewares/recovery"
-	"github.com/dadrus/heimdall/internal/handler/proxy/middlewares/trustedproxy"
+	"github.com/dadrus/heimdall/internal/handler/proxy/middleware/accesslog"
+	cachemiddleware "github.com/dadrus/heimdall/internal/handler/proxy/middleware/cache"
+	"github.com/dadrus/heimdall/internal/handler/proxy/middleware/dump"
+	"github.com/dadrus/heimdall/internal/handler/proxy/middleware/errorhandler"
+	"github.com/dadrus/heimdall/internal/handler/proxy/middleware/logger"
+	prometheusmiddleware "github.com/dadrus/heimdall/internal/handler/proxy/middleware/prometheus"
+	"github.com/dadrus/heimdall/internal/handler/proxy/middleware/recovery"
+	"github.com/dadrus/heimdall/internal/handler/proxy/middleware/trustedproxy"
 	"github.com/dadrus/heimdall/internal/heimdall"
 	"github.com/dadrus/heimdall/internal/rules/rule"
 	"github.com/dadrus/heimdall/internal/x"
